addeed missing } in utils, added login and signup components, added login component to login page, added values and onchange to signup inputs
